+++ src/utils/hooks/index.ts
@@ -1,5 +1,2 @@
 export * from "./useDispatch";
 export * from "./useQueries";
-
-// intentionally not exported
-// export * from "./useSelector";
