+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from "./containerUtils";
-export * from "./useSelector";
 export * from "./context";
 export * from "./hooks";
